+++ scrapers/scrapers/spiders/nv-bills.py
@@ -1,21 +1,201 @@
 import logging
 import re
 import time
+import dateutil.parser
+import pytz
 
-from scrapy import Request, Spider
+import requests
+
+from collections import defaultdict
 from dataclasses import dataclass
 
+from scrapy import Request, Selector, Spider
+
 from core.scrape import Bill
-from .consts import session_slugs, CARRYOVERS
 from .exceptions import BillTitleLengthError, SelectorError
-from .utils import (
-    add_actions,
-    add_sponsors,
-    extract_bdr,
-    get_column_div,
-    shorten_bill_title,
-)
 from ..items import BillStub, Chamber
+
+
+ACTION_CLASSIFIERS = (
+    ("Approved by the Governor", ["executive-signature"]),
+    ("Bill read. Veto not sustained", ["veto-override-passage"]),
+    ("Bill read. Veto sustained", ["veto-override-failure"]),
+    ("Enrolled and delivered to Governor", ["executive-receipt"]),
+    ("From committee: .+? adopted", ["committee-passage"]),
+    # the committee and chamber passage can be combined, see NV 80 SB 506
+    (
+        r"From committee: .+? pass(.*)Read Third time\.\s*Passed\.",
+        ["committee-passage", "reading-3", "passage"],
+    ),
+    ("From committee: .+? pass", ["committee-passage"]),
+    ("Prefiled. Referred", ["introduction", "referral-committee"]),
+    ("Read first time. Referred", ["reading-1", "referral-committee"]),
+    ("Read first time.", ["reading-1"]),
+    ("Read second time.", ["reading-2"]),
+    ("Read third time. Lost", ["failure", "reading-3"]),
+    ("Read third time. Passed", ["passage", "reading-3"]),
+    ("Read third time.", ["reading-3"]),
+    ("Rereferred", ["referral-committee"]),
+    ("Resolution read and adopted", ["passage"]),
+    ("Enrolled and delivered", ["enrolled"]),
+    ("To enrollment", ["passage"]),
+    ("Approved by the Governor", ["executive-signature"]),
+    ("Vetoed by the Governor", ["executive-veto"]),
+)
+
+session_slugs = {
+    "2010Special26": "26th2010Special",
+    "2013Special27": "27th2013Special",
+    "2014Special28": "28th2014Special",
+    "2015Special29": "29th2015Special",
+    "2016Special30": "30th2016Special",
+    "2020Special31": "31st2020Special",
+    "2020Special32": "32nd2020Special",
+    "2021Special33": "33rd2021Special",
+    "2023Special34": "34th2023Special",
+    "2023Special35": "35th2023Special",
+    "75": "75th2009",
+    "76": "76th2011",
+    "77": "77th2013",
+    "78": "78th2015",
+    "79": "79th2017",
+    "80": "80th2019",
+    "81": "81st2021",
+    "82": "82nd2023",
+}
+
+# NV sometimes carries-over bills from previous sessions,
+# without regard for bill number conflicts.
+# so AB1* could get carried in, even if there's already an existing AB1
+# The number of asterisks represent which past session it was pulled in from,
+# which can include specials, and skip around, so this can't be automated.
+# The list is at https://www.leg.state.nv.us/Session/81st2021/Reports/BillsListLegacy.cfm?DoctypeID=1
+# where 81st2021 will need to be swapped in for the session code.
+CARRYOVERS = {
+    "80": {
+        "*": "2017",
+    },
+    "81": {
+        "*": "2019",
+        "**": "2020Special32",
+    },
+    "82": {"*": "2021"},
+}
+
+
+TZ = pytz.timezone("PST8PDT")
+
+
+def get_column_div(name, response):
+    # lots of places where we have a <div class='col-md-2 font-weight-bold'>
+    # followeed by a <div class='col'>
+    # with interesting content in the latter element
+    return response.xpath(
+        f"//div[contains(text(),'{name}')]/following-sibling::div[@class='col']"
+    )
+
+
+def parse_date(date_str):
+    return TZ.localize(dateutil.parser.parse(date_str))
+
+
+def add_sponsors(bill, sponsor_links, primary):
+    seen = set()
+    for link in sponsor_links:
+        name = link.xpath('text()').get().strip()
+        if "Sponsors" in name or name == "":
+            continue
+        # Removes leg position from name
+        # Example: Assemblywoman Alexis Hansen
+        if name.split()[0] in ["Assemblywoman", "Assemblyman", "Senator"]:
+            name = " ".join(name.split()[1:]).strip()
+        if name not in seen:
+            seen.add(name)
+            bill.add_sponsorship(
+                name=name,
+                classification="sponsor" if primary else "cosponsor",
+                entity_type="person",
+                primary=primary,
+            )
+
+
+def add_actions(bill, chamber, response):
+    # first action is from originating chamber
+    actor = chamber
+
+    # Sometimes NV bill page might just not have an actions section at all
+    try:
+        for row in response.xpath("//caption/parent::table/tbody/tr"):
+            date = row.css('[data-th="Date"]::text').get()
+            action = row.css('[data-th="Action"]::text').get()
+
+            date = parse_date(date)
+
+            # catch chamber changes
+            if action.startswith("In Assembly"):
+                actor = "lower"
+            elif action.startswith("In Senate"):
+                actor = "upper"
+            elif "Governor" in action:
+                actor = "executive"
+
+            action_type = []
+            for pattern, atype in ACTION_CLASSIFIERS:
+                if not re.search(pattern, action, re.IGNORECASE):
+                    continue
+                # sometimes NV returns multiple actions in the same posting
+                # so don't break here
+                action_type = action_type + atype
+
+            if not action_type:
+                action_type = None
+            else:
+                action_type = list(set(action_type))
+
+            related_entities = []
+            if "Committee on" in action:
+                committees = re.findall(
+                    r"Committee on ([a-zA-Z, ]*)\.", action)
+                for committee in committees:
+                    related_entities.append(
+                        {"type": "committee", "name": committee}
+                    )
+
+            bill.add_action(
+                description=action,
+                date=date,
+                chamber=actor,
+                classification=action_type,
+                related_entities=related_entities,
+            )
+    except SelectorError:
+        pass
+
+
+def extract_bdr(title):
+    """
+    bills in NV start with a 'bill draft request'
+    the number is in the title but it's useful as a structured extra
+    """
+    bdr = False
+    bdr_regex = re.search(r"\(BDR (\w+-\w+)\)", title)
+    if bdr_regex:
+        bdr = bdr_regex.group(1)
+    return bdr
+
+
+def shorten_bill_title(title):
+    """
+    Used in cases where the bill title exceeds the
+    300-character limit that we have on this attribute.
+    """
+    title_bdr_re = re.compile(r"(.+)(\(BDR \w+-\w+\))")
+    title_bdr_match = title_bdr_re.search(title)
+    bdr_full = ""
+    if title_bdr_match:
+        title, bdr_full = title_bdr_match.groups()
+    title = f"{title[:280]}... + {bdr_full}"
+    return title
 
 
 @dataclass
@@ -29,17 +209,45 @@ class BillsSpider(Spider):
 
     def __init__(self, session=None, **kwargs):
         self.session = session
+        self.subject_mapping = defaultdict(set)
         super().__init__(**kwargs)
 
     def start_requests(self):
         # TODO default active sessions
         slug = session_slugs[self.session]
+        # subject_mapping
+        year = slug[4:]
+        url = (
+            f"https://www.leg.state.nv.us/Session/{slug}/Reports/"
+            f"TablesAndIndex/{year}_{self.session}-index.html"
+        )
+
+        dependency_request = requests.get(url)
+        dependency_response = Selector(dependency_request)
+        # first, a bit about this page:
+        # Level0 are the bolded titles
+        # Level1,2,3,4 are detailed titles, contain links to bills
+        # all links under a Level0 we can consider categorized by it
+        # there are random newlines *everywhere* that should get replaced
+        subject = None
+        for p in dependency_response.xpath("//p"):
+            if p.xpath("@class").get() == "Level0":
+                subject = p.xpath('text()').get().replace("\r\n", " ")
+            else:
+                if subject:
+                    for a in p.xpath(".//a"):
+                        bill_id = a.xpath('text()').get().replace(
+                            "\r\n", "") if a.xpath('text()').get() else None
+                        self.subject_mapping[bill_id].add(subject)
+
+        # proceed bill listing
         bill_listing_url = (f"https://www.leg.state.nv.us/App/NELIS/REL/{slug}/"
                             "HomeBill/BillsTab?selectedTab=List&Filters.PageSize=2147483647"
                             f"&_={time.time()}")
         meta = {
             "is_scout": True
         }
+
         yield Request(url=bill_listing_url, callback=self.parse_bill_list, meta=meta)
 
     def parse_bill_list(self, response):
@@ -55,7 +263,7 @@ class BillsSpider(Spider):
                 link.xpath("text()").get(),
                 self.session,
                 chamber,
-                []
+                list(self.subject_mapping[identifier])
             )
 
             # Yield the stub
